requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0.1
django-admin-interface==0.28.3
django-ckeditor==6.7.0
django-colorfield==0.11.0
django-js-asset==2.2.0
django-solo==2.1.0
idna==3.6
pillow==10.2.0
python-slugify==8.0.1
requests==2.31.0
sqlparse==0.4.4
stripe==7.11.0
text-unidecode==1.3
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
